+++ base/src/main/java/org/aya/concrete/remark/CodeAttrProcessor.java
@@ -1,6 +1,6 @@
 // Copyright (c) 2020-2023 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.concrete.remark.code;
+package org.aya.concrete.remark;
 
 import com.intellij.openapi.util.text.StringUtil;
 import kala.collection.mutable.MutableList;
