+++ base/src/main/java/org/aya/concrete/remark/CodeOptions.java
@@ -1,8 +1,7 @@
 // Copyright (c) 2020-2023 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.concrete.remark.code;
+package org.aya.concrete.remark;
 
-import org.aya.concrete.remark.AyaLiterate;
 import org.aya.generic.util.NormalizeMode;
 import org.aya.literate.Literate;
 import org.aya.pretty.doc.Doc;
